Almost done with shop
Gotta add in gold, need to tie shop in with gameplay, and need to switch back to play scene when continue is pressed

+++ Assets/Javyn/Scripts/Singleton.cs
@@ -7,7 +7,7 @@ public class Singleton : MonoBehaviour
 {
     public static Singleton singleton;
     
-    public int gold = 0;
+    public double gold = 0;
     public int depthUprage = 1;
     public int cashUpgrade = 1;
     public int spawnUpgrade = 1;
